Move functions into new file
