+++ src/game_logic/game_world.cpp
@@ -38,6 +38,7 @@
 #include "game_logic/enemies/dying_boss.hpp"
 #include "game_logic/world_state.hpp"
 #include "renderer/upscaling_utils.hpp"
+#include "renderer/viewport_utils.hpp"
 #include "ui/menu_element_renderer.hpp"
 #include "ui/utils.hpp"
